Google_Cloud_Function/requirements.txt
firebase-admin
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage